android/app/src/main/java/com/vava/mapp/EffectsManager.kt
package com.vava.mapp

import android.util.Log

class EffectsManager {

    fun loadEffectPreset(jsonPreset: String) {
        // Logique pour charger les presets d'effets depuis un fichier JSON
    }

    fun applyEffect(effectName: String) {
        // Applique l'effet spécifié
        Log.d("EffectsManager", "Applying effect: $effectName")
    }

    fun listAvailableEffects(): List<String> {
        // Liste les effets disponibles
        return listOf("Reverb", "Echo", "Bass Boost")
    }
}
